hcr-webapp/requirements.txt
fastapi
uvicorn
jinja2
biopython
python-multipart
pandas
openpyxl
numpy
pywebview
